+++ Source/Scene/GlobeSurfaceShaderSet.js
@@ -80,6 +80,7 @@ GlobeSurfaceShaderSet.prototype.getShaderProgram = function (options) {
   var applySaturation = options.applySaturation;
   var applyGamma = options.applyGamma;
   var applyAlpha = options.applyAlpha;
+  var applyDayNightAlpha = options.applyDayNightAlpha;
   var applySplit = options.applySplit;
   var showReflectiveOcean = options.showReflectiveOcean;
   var showOceanWaves = options.showOceanWaves;
@@ -155,7 +156,8 @@ GlobeSurfaceShaderSet.prototype.getShaderProgram = function (options) {
     (highlightFillTile << 24) |
     (colorToAlpha << 25) |
     (showUndergroundColor << 26) |
-    (translucent << 27);
+    (translucent << 27) |
+    (applyDayNightAlpha << 28);
 
   var currentClippingShaderState = 0;
   if (defined(clippingPlanes) && clippingPlanes.length > 0) {
@@ -220,6 +222,9 @@ GlobeSurfaceShaderSet.prototype.getShaderProgram = function (options) {
     }
     if (applyAlpha) {
       fs.defines.push("APPLY_ALPHA");
+    }
+    if (applyDayNightAlpha) {
+      fs.defines.push("APPLY_DAY_NIGHT_ALPHA");
     }
     if (showReflectiveOcean) {
       fs.defines.push("SHOW_REFLECTIVE_OCEAN");
@@ -290,7 +295,7 @@ GlobeSurfaceShaderSet.prototype.getShaderProgram = function (options) {
 
     var computeDayColor =
       "\
-    vec4 computeDayColor(vec4 initialColor, vec3 textureCoordinates)\n\
+    vec4 computeDayColor(vec4 initialColor, vec3 textureCoordinates, float nightBlend)\n\
     {\n\
         vec4 color = initialColor;\n";
 
@@ -333,6 +338,10 @@ GlobeSurfaceShaderSet.prototype.getShaderProgram = function (options) {
         (applyAlpha ? "u_dayTextureAlpha[" + i + "]" : "1.0") +
         ",\n\
             " +
+        (applyDayNightAlpha ? "u_dayTextureNightAlpha[" + i + "]" : "1.0") +
+        ",\n" +
+        (applyDayNightAlpha ? "u_dayTextureDayAlpha[" + i + "]" : "1.0") +
+        ",\n" +
         (applyBrightness ? "u_dayTextureBrightness[" + i + "]" : "0.0") +
         ",\n\
             " +
@@ -352,7 +361,8 @@ GlobeSurfaceShaderSet.prototype.getShaderProgram = function (options) {
         ",\n\
             " +
         (colorToAlpha ? "u_colorsToAlpha[" + i + "]" : "vec4(0.0)") +
-        "\n\
+        ",\n\
+        nightBlend\
         );\n";
       if (hasImageryLayerCutout) {
         computeDayColor +=